# swagger.yaml
openapi: 3.0.0
info:
  title: Directorio CSI API
  version: 1.0.0
  description: API para la gestión del directorio de contactos Grupo CSI
  contact:
    name: Tech-Potions
servers:
  - url: http://localhost:4000
    description: Desarrollo
  - url: https://freejolitos.com/apiQR
    description: Producción

paths:

  # ── Autenticación ────────────────────────────────────────────────────────────

  /accounts/authenticate:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      description: Autentica con email y contraseña. La cuenta debe estar verificada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: admin@grupocsi.com
                password:
                  type: string
                  example: miContraseña123
      responses:
        '200':
          description: Login exitoso — devuelve JWT y refreshToken en cookie
          headers:
            Set-Cookie:
              description: refreshToken (HTTP-only, 7 días)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountWithToken'
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts/refresh-token:
    post:
      tags: [Auth]
      summary: Renovar access token
      description: Usa el refreshToken de la cookie para emitir un nuevo JWT.
      responses:
        '200':
          description: Token renovado
          headers:
            Set-Cookie:
              description: Nuevo refreshToken (HTTP-only)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountWithToken'
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts/revoke-token:
    post:
      tags: [Auth]
      summary: Cerrar sesión (revocar refresh token)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: Si se omite, se usa el token de la cookie
      responses:
        '200':
          description: Token revocado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ── Verificación y recuperación de contraseña ────────────────────────────────

  /accounts/verify-email:
    post:
      tags: [Auth]
      summary: Verificar email
      description: Activa la cuenta con el token recibido por correo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  example: 3c7f8d9c4cb348ff95a0b74a1452aa24fc9611bb
      responses:
        '200':
          description: Email verificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts/forgot-password:
    post:
      tags: [Auth]
      summary: Solicitar recuperación de contraseña
      description: Envía un email con enlace de restablecimiento (válido 24 h).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  example: usuario@grupocsi.com
      responses:
        '200':
          description: Email enviado (responde igual aunque el email no exista)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /accounts/validate-reset-token:
    post:
      tags: [Auth]
      summary: Validar token de recuperación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts/reset-password:
    post:
      tags: [Auth]
      summary: Restablecer contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password, confirmPassword]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 6
                  example: nuevaContraseña123
                confirmPassword:
                  type: string
                  example: nuevaContraseña123
      responses:
        '200':
          description: Contraseña restablecida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Usuarios (solo Admin) ─────────────────────────────────────────────────────

  /accounts:
    get:
      tags: [Usuarios]
      summary: Listar usuarios
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de todos los usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags: [Usuarios]
      summary: Crear usuario
      description: Solo Admin. Envía email de verificación al nuevo usuario.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountInput'
      responses:
        '200':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /accounts/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        example: 1

    get:
      tags: [Usuarios]
      summary: Obtener usuario por ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Datos del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Usuarios]
      summary: Actualizar usuario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateInput'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [Usuarios]
      summary: Eliminar usuario
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuario eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ── VCards (público + Admin) ──────────────────────────────────────────────────

  /vcard/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        example: 5

    get:
      tags: [VCards]
      summary: Obtener vCard por ID (público)
      description: >
        Endpoint público usado al escanear el QR.
        Devuelve todos los campos incluyendo foto (base64) y vcf.
      responses:
        '200':
          description: Datos de la vCard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VCard'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /vcard:
    get:
      tags: [VCards]
      summary: Listar vCards
      description: Solo Admin. Excluye foto y vcf por rendimiento.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de vCards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VCardSummary'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags: [VCards]
      summary: Crear vCard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VCardInput'
      responses:
        '201':
          description: VCard creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VCard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /vcard/{id}/put:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        example: 5

    put:
      tags: [VCards]
      summary: Actualizar vCard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VCardInput'
      responses:
        '200':
          description: VCard actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VCard'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [VCards]
      summary: Eliminar vCard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: VCard eliminada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

# ── Componentes reutilizables ─────────────────────────────────────────────────

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    Account:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombres:
          type: string
          example: Juan
        primerApellido:
          type: string
          example: Pérez
        segundoApellido:
          type: string
          example: García
        email:
          type: string
          example: juan.perez@grupocsi.com
        role:
          type: string
          enum: [Admin, User]
          example: Admin
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        isVerified:
          type: boolean
          example: true

    AccountWithToken:
      allOf:
        - $ref: '#/components/schemas/Account'
        - type: object
          properties:
            jwtToken:
              type: string
              description: JWT de acceso (expira en 15 min)

    AccountInput:
      type: object
      required: [nombres, primerApellido, email, password, confirmPassword, role]
      properties:
        nombres:
          type: string
          example: Juan
        primerApellido:
          type: string
          example: Pérez
        segundoApellido:
          type: string
          example: García
        email:
          type: string
          example: juan.perez@grupocsi.com
        password:
          type: string
          minLength: 6
          example: contraseña123
        confirmPassword:
          type: string
          example: contraseña123
        role:
          type: string
          enum: [Admin, User]
          example: User

    AccountUpdateInput:
      type: object
      properties:
        nombres:
          type: string
        primerApellido:
          type: string
        segundoApellido:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
        role:
          type: string
          enum: [Admin, User]

    VCard:
      type: object
      properties:
        contactID:
          type: integer
          example: 5
        nombres:
          type: string
          example: María
        primerApellido:
          type: string
          example: López
        segundoApellido:
          type: string
          example: Sánchez
        empresa:
          type: string
          example: Grupo CSI
        puesto:
          type: string
          example: Directora Comercial
        email:
          type: string
          example: maria.lopez@grupocsi.com
        telefonoMovil:
          type: string
          example: '+52 55 1234 5678'
        telefonoOficina:
          type: string
          example: '+52 55 8765 4321'
        Address:
          type: string
          example: Av. Reforma 123, CDMX
        LinkedInURL:
          type: string
          example: 'https://linkedin.com/in/marialopez'
        foto:
          type: string
          description: Imagen en base64 o URL
        vcf:
          type: string
          description: Archivo vCard (.vcf) en base64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    VCardSummary:
      description: VCard sin foto ni vcf (para listados)
      type: object
      properties:
        contactID:
          type: integer
        nombres:
          type: string
        primerApellido:
          type: string
        segundoApellido:
          type: string
        empresa:
          type: string
        puesto:
          type: string
        email:
          type: string
        telefonoMovil:
          type: string
        telefonoOficina:
          type: string
        Address:
          type: string
        LinkedInURL:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    VCardInput:
      type: object
      required: [nombres, primerApellido, email, telefonoMovil]
      properties:
        nombres:
          type: string
          example: María
        primerApellido:
          type: string
          example: López
        segundoApellido:
          type: string
          example: Sánchez
        empresa:
          type: string
          example: Grupo CSI
        puesto:
          type: string
          example: Directora Comercial
        email:
          type: string
          example: maria.lopez@grupocsi.com
        telefonoMovil:
          type: string
          example: '+52 55 1234 5678'
        telefonoOficina:
          type: string
        Address:
          type: string
        LinkedInURL:
          type: string
        foto:
          type: string
          description: Base64 o URL
        vcf:
          type: string
          description: Base64 del archivo .vcf

    Message:
      type: object
      properties:
        message:
          type: string
          example: Operación exitosa

  responses:
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    UnauthorizedError:
      description: Token faltante, inválido o sin permisos suficientes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'